Fix error with FindResultsView attachment and update on react editor
When you’d split-open a file, the react editor wasn’t attached or 
rendered yet. So there was no under layer, and when you’d try to request
a display update, it would call functions on objects that weren’t there.

So I hack this with a delay. This is not a pretty fix. It will all go 
away when there are marker views in core.

refs #232