Transfer focus to resultsView subview on ResultsPaneView focus events
Previously, we transferred focus by overriding the ::focus method on
the results view and then calling focus on a subview. The outer view
itself wasn’t actually focusable. However, now the pane system calls
focus on the underlying DOM node, not the wrapper SpacePen view. This
is in preparation to deprecate SpacePen entirely. In order to be
compatible, the view itself now needs to be focusable and we transfer
focus via the event.