Make project search pluggable via DefaultSearchDirectoryProvider.
This refactors `ResultsModel` so that when its `search()` method is
invoked, it finds the appropriate `SearchDirectoryProvider` for each
`Directory` in `atom.project.getDirectories()`. It invokes the `search()`
method for each provider and aggregates the results as they are streamed in.

Currently, there is only one implementation of `SearchDirectoryProvider`:
`DefaultSearchDirectoryProvider`. It uses the existing logic from
`ResultsModel` to implement `search()`, which is basically a call to
`atom.workspace.scan()` with an appropriate callback.

The next step is to make `find-and-replace` a consumer of `SearchDirectoryProvider`
objects so that remote directories can be supported. New providers will be added
to the front of the `ResultsModel::searchProviders` array so that they are considered
before the `DefaultSearchDirectoryProvider`. Note that the same technique is used
with `Project::directoryProviders`.