Prevent search results from being ignored in specs
As of atom/scandal#40, VCS ignores are respected even when
the the project path is a subdirectory of the repository path.
This means that if the tests are run from the `atom/node_modules`
folder, atom's .gitignore will cause all results in the package
directory to be ignored. Thus, we disable `excludeVcsIgnores` in
the specs.