Fix deceptive class when an error occurs
Consider the following scenario:
1. Turn on regex
2. Search for something simple in the file, eg "this"
3. has-results class appears as expected
4. Add an opening parenthesis to the find pattern
5. Find box is still green?!?!
Step 5 introduces contradicting information: a red error message yet a
green find box with the has-results class.

So now when we set an error message we make sure to remove the
has-results class and add the has-no-results class.

I'm also fairly certain that nulling a function is not a good idea, so I
removed that line.