Don't show find-and-replace "Copy Path" item in tree-view context menu
befebde9a6830e126a156cb33b5a8b99a77191d6 reduced the specificity of the
selector for the "Copy Path" menu item [A], causing the menu item to
appear in the tree-view's context menu. With the changes in this commit,
the "Copy Path" menu item appears in the find-and-replace context menu
for files and lines, but it does not appear in the tree-view context
menu.

[A]
https://github.com/atom/find-and-replace/commit/befebde9a6830e126a156cb33b5a8b99a77191d6#diff-9d7d0803c9fece359b6cd239f42a2e23L31